[cherry-pick] fix(iota-node): unregister and re-register consensus adapter metrics during reconfiguration (#6283) (#6288)
# Description of change

This PR fixes an issue with an unwrap panic during creation of
ConsensusAdapterMetrics with error AlreadyReg. Metrics are created
during reconfiguration when a node becomes a validator. It a node
becomes a validator for the second time (after eg. "leaving validator
set") the metrics will be created for the second time reusing the same
"default" registry which would lead to a crash as metrics are never
unregistered.

This PR adds ConsensusAdapterMetrics::unregister function to unregister
previously created metrics (it can't be done in `drop` as we need the
registry). It is called when a node is no longer a validator after
reconfiguration. HistogramVec from iota-metrics counters are dropped
separately.

## Links to any relevant issues

fix #6277

## Type of change

- Bug fix (a non-breaking change which fixes an issue)


## Change checklist

Tick the boxes that are relevant to your changes, and delete any items
that are not.

- [ ] I have followed the contribution guidelines for this project
- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes

---------

# Description of change

Please write a summary of your changes and why you made them.

## Links to any relevant issues

Be sure to reference any related issues by adding `fixes #(issue)`.

## Type of change

Choose a type of change, and delete any options that are not relevant.

- Bug fix (a non-breaking change which fixes an issue)
- Enhancement (a non-breaking change which adds functionality)
- Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- Documentation Fix

## How the change has been tested

Describe the tests that you ran to verify your changes.

Make sure to provide instructions for the maintainer as well as any
relevant configurations.

## Change checklist

Tick the boxes that are relevant to your changes, and delete any items
that are not.

- [ ] I have followed the contribution guidelines for this project
- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes

Co-authored-by: jkrvivian <jkrvivian@gmail.com>