upstream(node): Add salt to TrafficSketch (#7445)
# Description of change

- Upstream range: [v1.37.4, v1.38.4)
- Port commit:
https://github.com/MystenLabs/sui/commit/53dd080a841710c91f492de17f7312a29662235f
- Descriptions from commits
False positives stem from IP collisions during traffic tallying. Because
validators tend to see closely correlated transaction execution traffic
patterns, it is possible that a false positive resulting in an
erroneously blocked IP could be correlated across many validators at the
same time. If this happens on a quorum of validators, this would result
in an IP being effectively denied service until the TTL expires.

Luckily, we can exploit this property to lower the effective false
positive rate (as observed by the client) by removing this correlation.
This PR adds a salt key to the `SketchKey`. The value is generated in
memory at startup.


## Links to any relevant issues

Part of #4390  

## How the change has been tested

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes