fix(docs): change `iotaleger` to `iotaledger` (#7562)
# Description of change

This change fixes a broken link due to a typo

## Links to any relevant issues

n/a

## How the change has been tested

Clicking the link works now

## Change checklist

Tick the boxes that are relevant to your changes, and delete any items
that are not.

- [x] I have followed the [contribution
guidelines](../../CONTRIBUTING.md) for this project
- [x] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation

### Release Notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK:
- [ ] REST API:

Co-authored-by: Thibault Martinez <thibault@iota.org>
Co-authored-by: Dr-Electron <dr-electr0n@protonmail.com>