feat(ci): reduce amount of benchmark processes running in parallel in the nightlies (#8032)
# Description of change

This PR reduces the amount of benchmark processes running in parallel in
the nightlies. Before they were only limited by "NUM_CPU", which is 32
in our backend. I guess we can create more load in those tests, if we
actually run less workers.

Also I fixed a confusing log level, in case the safe approach for epoch
database deletion did fail. It was changed from "error" to "warn",
because there is a fallback mechanism in place in case this happens.

## How the change has been tested

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes