fix(iota-core): use current commit suggested gas price instead of `expect` for previously deferred txs (#8035)
# Description of change

It appears to be possible that a transaction deferred due to randomness
unavailable in a previous commit round can be deferred due to congestion
in the current commit round. That is, such a transaction will appear in
the list of previously deferred transactions, but the suggested gas
price for transactions deferred due to randomness unavailable is set to
`None`. Then, in the current round, we will not be able to get a
suggested gas price from the previous commit round. In such case,
instead of `expect` (panic), we will return the suggested gas price from
the current commit round. This is both safer and fair, as the current
commit round appears the first one in which the transaction is deferred
due to congestion.

## Links to any relevant issues

## How the change has been tested

```console
MSIM_TEST_SEED=211753826735 cargo simtest -p iota-benchmark test_simulated_load_shared_object_congestion_control
```

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [x] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have checked that new and existing unit tests pass locally with
my changes

### Release Notes

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK:
- [ ] REST API: