fix(graphql): increase gas coin amount for dry runs (#8444)
# Description of change

Increase the gas coin amount for the created gas coin that's used in dry
runs with graphql. It was just 1_000 IOTA and therefore caused issues
with insufficient funds. Now it's 1_000_000_000 IOTA, the same value as
used for dry runs and more than any account controls in the mainnet,
this should be sufficient large. Using something like u64::MAX -1
doesn't work, because if there are then other coins in a transaction,
this would exceed u64 and result in an overflow error.

## Links to any relevant issues

Fixes https://github.com/iotaledger/iota/issues/8440

## How the change has been tested

- [ ] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes

### Release Notes

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [x] GraphQL: Increased dry run amount from 1_000 to 1_000_000_000 IOTA
- [ ] CLI:
- [ ] Rust SDK:
- [ ] REST API: