feat(json-rpc): add new method isTransactionIndexedOnNode (#8593)
# Description of change

This patch adds a new endpoint `iota_isTransactionIndexedOnNode`

## Links to any relevant issues

Fixes #8526

## How the change has been tested

Describe the tests that you ran to verify your changes.

Make sure to provide instructions for the maintainer as well as any
relevant configurations.

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [x] Patch-specific tests (correctness, functionality coverage)

### Infrastructure QA (only required for crates that are maintained by
@iotaledger/infrastructure)

- [ ] Synchronization of the indexer from genesis for a network
including migration objects.
- [ ] Restart of indexer synchronization locally without resetting the
database.
- [ ] Restart of indexer synchronization on a production-like database.
- [ ] Deployment of services using Docker.
- [x] Verification of API backward compatibility.

### Release Notes

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [x] JSON-RPC: A new method `iota_isTransactionIndexedOnNode` is added
that allows clients and dapps to optionally poll the fullnode to verify
that a transaction has been indexed. The latter is a necessary condition
so that any dependent transaction is executed successfully. The new
method complements the recent advancements in the Indexer JSON-RPC API
that indexes transactions after execution without waiting for a
checkpoint to include them.
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK:
- [ ] REST API: