fix(framework-snapshot): Solve git fetch shallowness in check_manifest_against_tomls test (#8644)
# Description of change

This PR tries to fix the `fatal: internal error in revision walk
failure` thrown occasionally in the CI when the
`check_manifest_against_tomls` test is executed. The solution to this
fix should be adding `--depth=2147483647` to the git fetch command
executed during the test.

Moreover, an additional check makes so that this test is executed only
then the `iota-framework-snapshot/manifest.json` file is modified.

## Links to any relevant issues

Fixes #7973 

## How the change has been tested

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [x] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have checked that new and existing unit tests pass locally with
my changes