refactor(indexer): insert object deletion and mutation chunks concurrently (#8723)
# Description of change

This patch makes a minor optimization in the iota-indexer following the
respective upstream changes.

## Links to any relevant issues

Resolves #8676

## How the change has been tested

Describe the tests that you ran to verify your changes.

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [x] Patch-specific tests (correctness, functionality coverage)

### Infrastructure QA (only required for crates that are maintained by
@iotaledger/infrastructure)

No further QA tests were deemed necessary as this patch only makes the
insertion of deletion and mutations concurrent, on the guarantee that
all objects are unique and have the highest version among a set.