fix(consensus): reading own blocks returns only fully available blocks from DAG state (#8876)
# Description of change

Due to the non-atomic nature of storing transactions and headers in the
DAG state, we could have an own block header, but no transactions for
that header leading to the panic when calling method
`get_own_cached_blocks`. In particular, that is used after getting a
subscription for streaming own blocks.

We change the method `get_own_cached_blocks` to return only blocks which
have both headers and transactions and don't return those with missing
parts (e.g. transactions).

## Links to any relevant issues

Fixes #8874

## How the change has been tested

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [x] I have checked that new and existing unit tests pass locally with
my changes

### Release Notes

---------

Co-authored-by: Piotr Macek <4007944+piotrm50@users.noreply.github.com>