feat(starfish): send only useful shards (#8801)
# Description of change

PR tracks shards of which authors could be sent to other validators.

Shards for encoded transaction data of validator $v$ are considered as
useful for validator $u$ if validator $u$ is ready to include an
additional block header of validator $v$ after filtering those blocks.
In particular, this means that
the potential sequenced transaction data comes from validator $v$
indirectly.

If the shard of the block created by the validator $v$ was useful, then
the shards from the blocks created by $v$ will be sent for some number
of rounds. In other words, there is some inertia; we don't stop sending
shards as soon as one of them is useless.

In practice, there are two prerequisites for the efficiency of this
mechanism:
1) block bundle streaming is actively working (e.g. this mechanism
wouldn't work if a validator is catching up and doesn't produce blocks)
2) cordial dissemination of block headers is enabled (without additional
block headers, validators won't understand which shards are useful for
them)

## Links to any relevant issues

fixes #8767 

## How the change has been tested

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have checked that new and existing unit tests pass locally with
my changes