fix(graphql-rpc): read also checkpointed transactions after execution (#8935)
# Description of change

Fixes `mutation.executeTransactionBlock` which could fail while trying
to construct the transaction response. The reason being that it was
expecting that the transaction would be indexed through the optimistic
path, which is not necessarily true.

## How the change has been tested

Describe the tests that you ran to verify your changes.

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)

### Release Notes

- [x] GraphQL: Fixes `mutation.executeTransactionBlock` which could fail
if the transaction executed was indexed through the checkpoint path.

---------

Co-authored-by: Sergiu Popescu <44298302+sergiupopescu199@users.noreply.github.com>