upstream(iota): fix `iota move build --dump-bytecode-as-base64` to correctly take into account Move.lock addresses (#8931)
# Description of change

This PR fixes a bug where `iota move build --dump-bytecode-as-base64`
uses a published address in the resolved table for the root package
being built instead of 0x0, leading to inconsistencies when this
interacts with automated feature management.

## Links to any relevant issues

Based on https://github.com/MystenLabs/sui/pull/22528

## How the change has been tested

- [ ] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes

### Release Notes

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [x] CLI: `iota move build --dump-bytecode-as-base64` is now working
correctly due to a bug in the logic when a Move.lock file existed with
the published address.
- [ ] Rust SDK:
- [ ] REST API: