upstream(graphql): add mutation payload size limit (#9152)
# Description of change

This patch introduces a new limit specific to transaction execution
paths (`{mutation.execute,query.DryRun}TransactionBlock`. This allows
transaction payloads to exceed the existing query limit size (5KB) and
applies an upper that complies with the max size allowed by the
protocol.

This fixes issues with large transactions (typically `publish`) not
being able to execute.

`query.moveViewCall` is not handled here.

## Links to any relevant issues

Fixes #8726 

## How the change has been tested

Describe the tests that you ran to verify your changes.

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [x] Patch-specific tests (correctness, functionality coverage)

### Infrastructure QA (only required for crates that are maintained by
@iotaledger/infrastructure)

- [x] Verification of API backward compatibility.

### Release Notes

- [x] GraphQL: Adds a new `maxTxPayloadSize` configuration parameter
under the `limits` section. This is the maximum size in bytes allowed
for the `txBytes` and `signatures` payload in
`mutation.executeTransactionBlock` and for `txBytes` in
`query.dryRunTransactionBlock`. The default value is the max transaction
size allowed by the protocol (128KB) plus the `Base64` encoding overhead
(~1/3 of the size).