fix(consensus): track failures to choose optimal peers and adjust parameters in the Starfish transaction synchronizer (#9183)
# Description of change

Transaction synchronizer fetches transactions either slowly or with some
warning in many scenarios:
- after commit syncer has synced a big number of certified commits, the
transactions from the committed subDAGs are fetched quite slowly.
- when some connections are blocked, the transaction synchronizer
produces too many failures with blocked peers.

In this PR, 
- we tune parameters of the transaction synchronizer to increase the
number of concurrent high-level task and control the total number of
spawned tasks.
- we introduce a simple memory tracking when the transaction fetch has
failed for the last time for each given peer. This memory is then used
to take proper peers for fetching
- we remove transaction verification from the transaction synchronizer
(similarly to how it was done in the commit syncer) since the
verification and validity of transactions was already performed by a
quorum of the network.

## Links to any relevant issues

Fixes #9162 

## How the change has been tested
We have tested the change using the fuzz testing scenarios.

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have checked that new and existing unit tests pass locally with
my changes