fix(graphql-rpc)!: Change type of "iotaTotalSupply" to BigInt (#9289)
# Description of change

Since the value of "iotaTotalSupply" is too big to fit in Int or UInt53
Graphql types, we need to change it to "BigInt".
This will change the data type in the JSON payload returned by graphql
from int to string, so this is a breaking change.

## Links to any relevant issues

fixes https://github.com/iotaledger/iota/issues/9288

## How the change has been tested

Describe the tests that you ran to verify your changes.

Make sure to provide instructions for the maintainer as well as any
relevant configurations.

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)

### Infrastructure QA (only required for crates that are maintained by
@iotaledger/infrastructure)

- [ ] Synchronization of the indexer from genesis for a network
including migration objects.
- [ ] Restart of indexer synchronization locally without resetting the
database.
- [ ] Restart of indexer synchronization on a production-like database.
- [ ] Deployment of services using Docker.
- [ ] Verification of API backward compatibility.

### Release Notes

- [x] GraphQL: Type of `iotaTotalSupply` changed from `int` to `BigInt`