feat(aws): add `iota-build-cache-server` crate (#9423)
# Description of change

By using the `aws-orchestrator`, we spawn a lot of different instances
which first need to compile the different binaries used in the following
tests. This wastes around 20 minutes per instance. For bigger tests with
a lot of instances it makes sense to pre-compile the used binaries,
before actually starting all instances.

This PR introduces an `iota-build-cache-server`, which can be run on a
dedicated server, outside of the `aws-orchestrator` setup. The server
can pre-build the binaries using the additional bash scripts. After the
binaries are ready, the `aws-orchestrator` can be started and run tests
right away.

To use the `build-cache-server` in `aws-orchestrator`, this section
needs to be added to the orchestrator config:
```json
    "build_cache": {
        "enabled": true,
        "servers": {
            "x86_server": {
                "targets": ["x86-64", "x86-64-v2", "x86-64-v3"],
                "url": "http://localhost:8080",
                "username": "",
                "password": ""
            },
            "arm_server": {
                "targets": ["aarch64"],
                "url": "http://localhost:8081",
                "username": "",
                "password": ""
            }
        }
    },
```

## How the change has been tested

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes