feat(aws+swarm): improvements for spam via dedicated clients in orchestrator (#9631)
# Description of change

Currently the dedicated clients share the gas object from the genesis
with the validators. If there are more dedicated clients than
validators, the `stress` in `iota-benchmark` fails because the gas
objects get locked.

Also we need to clone the git repo in the orchestrator even if we are
using the build-cache, because we need it for grafana and the
iota-benchmark (move-examples and framework).

## How the change has been tested

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes