feat(graphql-rpc): Optimize reading objects_history in consistent views (#9563)
# Description of change

Optimize reading objects_history in consistent views.

The original approach was reading all matching (that passed `filter_fn`
and the `range` check) rows from `objects_history` and then filtered
them to leave only those that for given object_id DON'T have any other
row in `objects_history` that satisfies `newer_criteria`, `range` check
and that have higher version than selected row.

The filter we use when getting objects is different than the filter we
use when trying to determine if it is the most recent version, and this
actually is relevant.

Imagine we have object A which at version 3 belongs to Bob, and at
Version 5 belongs to Alice.
Now if we want to find all objects belonging to Bob, and in our
available range we see A both in versions 3 and 5, we shouldn't return
A:3 because that's the most recent version that satisfies the criteria
but rather we should not return the object at all, because at the most
recent version of it in our available range the object doesn't belong to
Bob.

So the behaviour of the new query is like: "fetch all objects that match
the criteria, but eventually return only those that happen to be the
most recent version of given object in given checkpoint range", which
should give the same results as the original approach. (the assumption
for this is that `newer_criteria` will NOT filter out object version
that should be returned, and it is the case in current codebase).

## Links to any relevant issues

fixes https://github.com/iotaledger/iota/issues/7296

## How the change has been tested

Describe the tests that you ran to verify your changes.

Make sure to provide instructions for the maintainer as well as any
relevant configurations.

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)

### Infrastructure QA (only required for crates that are maintained by
@iotaledger/infrastructure)

- [ ] Synchronization of the indexer from genesis for a network
including migration objects.
- [ ] Restart of indexer synchronization locally without resetting the
database.
- [ ] Restart of indexer synchronization on a production-like database.
- [ ] Deployment of services using Docker.
- [ ] Verification of API backward compatibility.